feat: Apply git pull and push before each passwordstore playbook. Adapt to Ansible 2.10
At the start and finish of the playbooks that use/update the passwordstore a git pull (at the beggining) and a git push (at the end) is being executed to force synchronization of the git repository. Since the tests are being made with Ansible 2.10 the playbooks have been adapted to this version requiring the removal of the 'when' parameter of the 'vars_prompt'. One step further, so previous versions aren't affected, 'vars_prompt' have been replaced by the verification of all the 'vars_prompt' parameters with an assert validating them.